refactor(backend): restructure Application.kt into modular architecture
- Split monolithic Application.kt (740 lines) into proper package structure:
  * controller/ - REST API controllers
  * service/ - Business logic services
  * dto/ - Data transfer objects
  * tools/ - Koog agent tools (6 specialized classes)
  * config/ - Configuration classes
- Remove OpenAPI/validation dependencies for simplified setup
- Improve Japanese weather query handling with location mapping fallback
- Follow single responsibility principle for better maintainability
